pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Gmail MCP Server for sending emails via Gmail API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.3",
    "google-auth==2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-api-python-client>=2.177.0",
    "pydantic>=2.11.7",
    "cryptography>=45.0.6",
    "click>=8.2.1",
    "ruff>=0.12.7",
]

[project.scripts]
gmail-mcp = "main:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.black]
line-length = 88
target-version = ['py38']
